requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-MySQLdb==1.0.1
Flask-WTF==1.1.1
mysql-connector-python==8.1.0
Werkzeug==2.3.7
WTForms==3.0.1
PyMySQL==1.1.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9

python-dotenv==1.0.0 